Another big batch of small refactors.
